move ProfilePage folder into the Auth directory
