+++ src/components/Auth/ProfilePage/ProfileComponents/QR/Qr.js
